+++ filament/src/fg2/details/PassNode.h
@@ -21,7 +21,10 @@
 #include "fg2/details/Utilities.h"
 #include "fg2/FrameGraph.h"
 #include "fg2/FrameGraphRenderPass.h"
+
 #include "private/backend/DriverApiForward.h"
+
+#include <backend/TargetBufferInfo.h>
 
 #include <unordered_set>
 
@@ -63,13 +66,14 @@ class RenderPassNode : public PassNode {
 public:
     class RenderPassData {
     public:
+        static constexpr size_t ATTACHMENT_COUNT = backend::MRT::MAX_SUPPORTED_RENDER_TARGET_COUNT + 2;
         const char* name = {};
         FrameGraphRenderPass::Descriptor descriptor;
         bool imported = false;
         backend::TargetBufferFlags targetBufferFlags = {};
-        FrameGraphId<FrameGraphTexture> attachmentInfo[6] = {};
-        ResourceNode* incoming[6] = {};  // nodes of the incoming attachments
-        ResourceNode* outgoing[6] = {};  // nodes of the outgoing attachments
+        FrameGraphId<FrameGraphTexture> attachmentInfo[ATTACHMENT_COUNT] = {};
+        ResourceNode* incoming[ATTACHMENT_COUNT] = {};  // nodes of the incoming attachments
+        ResourceNode* outgoing[ATTACHMENT_COUNT] = {};  // nodes of the outgoing attachments
         struct {
             backend::Handle<backend::HwRenderTarget> target;
             backend::RenderPassParams params;