+++ libs/gltfio/src/ResourceLoader.cpp
@@ -17,6 +17,7 @@
 #include <gltfio/ResourceLoader.h>
 #include <gltfio/Image.h>
 
+#include "GltfEnums.h"
 #include "FFilamentAsset.h"
 #include "TangentsJob.h"
 #include "upcast.h"
@@ -27,6 +28,8 @@
 #include <filament/MaterialInstance.h>
 #include <filament/Texture.h>
 #include <filament/VertexBuffer.h>
+
+#include <geometry/Transcoder.h>
 
 #include <utils/JobSystem.h>
 #include <utils/Log.h>
@@ -52,6 +55,9 @@
 using namespace filament;
 using namespace filament::math;
 using namespace utils;
+
+using filament::geometry::Transcoder;
+using filament::geometry::ComponentType;
 
 static const auto FREE_CALLBACK = [](void* mem, size_t, void*) { free(mem); };
 
@@ -185,6 +191,33 @@ static void convertBytesToShorts(uint16_t* dst, const uint8_t* src, size_t count
     }
 }
 
+static ComponentType getComponentType(const cgltf_accessor* accessor) {
+    switch (accessor->component_type) {
+        case cgltf_component_type_r_8: return ComponentType::BYTE;
+        case cgltf_component_type_r_8u: return ComponentType::UBYTE;
+        case cgltf_component_type_r_16: return ComponentType::SHORT;
+        case cgltf_component_type_r_16u: return ComponentType::USHORT;
+        default:
+            // This should be unreachable because other types do not require conversion.
+            assert_invariant(false);
+            return {};
+    }
+}
+
+static void convertToFloats(float* dest, const cgltf_accessor* accessor) {
+    const uint32_t dim = cgltf_num_components(accessor->type);
+    const size_t floatsSize = accessor->count * sizeof(float) * dim;
+    Transcoder transcode({
+        .componentType = getComponentType(accessor),
+        .normalized = bool(accessor->normalized),
+        .componentCount = dim,
+        .inputStrideBytes = uint32_t(accessor->stride)
+    });
+    auto bufferData = (const uint8_t*) accessor->buffer_view->buffer->data;
+    const uint8_t* source = computeBindingOffset(accessor) + bufferData;
+    transcode(dest, source, accessor->count);
+}
+
 static void decodeDracoMeshes(FFilamentAsset* asset) {
     DracoCache* dracoCache = &asset->mSourceAsset->dracoCache;
 
@@ -200,25 +233,30 @@ static void decodeDracoMeshes(FFilamentAsset* asset) {
     };
 
     // Go through every primitive and check if it has a Draco mesh.
-    for (auto pair : asset->mPrimitives) {
+    for (auto& pair : asset->mPrimitives) {
         const cgltf_primitive* prim = pair.first;
-        VertexBuffer* vb = pair.second;
         if (!prim->has_draco_mesh_compression) {
             continue;
         }
 
         const cgltf_draco_mesh_compression& draco = prim->draco_mesh_compression;
 
+        // If an error occurs, we can simply set the primitive's associated VertexBuffer to null.
+        // This does not cause a leak because it is a weak reference.
+        auto& vertexBuffer = pair.second;
+
         // Check if we have already decoded this mesh.
         DracoMesh* mesh = dracoCache->findOrCreateMesh(draco.buffer_view);
         if (!mesh) {
-            slog.w << "Cannot decompress mesh, Draco decoding error." << io::endl;
+            slog.e << "Cannot decompress mesh, Draco decoding error." << io::endl;
+            vertexBuffer = nullptr;
             continue;
         }
 
         // Copy over the decompressed data, converting the data type if necessary.
-        if (prim->indices) {
-            mesh->getFaceIndices(prim->indices);
+        if (prim->indices && !mesh->getFaceIndices(prim->indices)) {
+            vertexBuffer = nullptr;
+            continue;
         }
 
         // Go through each attribute in the decompressed mesh.
@@ -237,7 +275,10 @@ static void decodeDracoMeshes(FFilamentAsset* asset) {
             }
 
             // Copy over the decompressed data, converting the data type if necessary.
-            mesh->getVertexAttributes(id, accessor);
+            if (!mesh->getVertexAttributes(id, accessor)) {
+                vertexBuffer = nullptr;
+                break;
+            }
         }
     }
 }
@@ -432,6 +473,17 @@ bool ResourceLoader::loadResources(FFilamentAsset* asset, bool async) {
         const uint8_t* data = computeBindingOffset(accessor) + bufferData;
         const uint32_t size = computeBindingSize(accessor);
         if (slot.vertexBuffer) {
+            if (requiresConversion(accessor->type, accessor->component_type)) {
+                const size_t dim = cgltf_num_components(accessor->type);
+                const size_t floatsSize = accessor->count * sizeof(float) * dim;
+                float* floatsData = (float*) malloc(floatsSize);
+                convertToFloats(floatsData, accessor);
+                BufferObject* bo = BufferObject::Builder().size(floatsSize).build(engine);
+                asset->mBufferObjects.push_back(bo);
+                bo->setBuffer(engine, BufferDescriptor(floatsData, floatsSize, FREE_CALLBACK));
+                slot.vertexBuffer->setBufferObjectAt(engine, slot.bufferIndex, bo);
+                continue;
+            }
             BufferObject* bo = BufferObject::Builder().size(size).build(engine);
             asset->mBufferObjects.push_back(bo);
             bo->setBuffer(engine, BufferDescriptor(data, size,